+++ bluer_sbc/.abcli/alias.sh
@@ -1,5 +1,7 @@
 #! /usr/bin/env bash
 
-alias @sbc=bluer_sbc
+alias @camera=bluer_sbc_camera
 
 alias grove=bluer_sbc_grove
+
+alias @sbc=bluer_sbc
