+++ bluer_sbc/__init__.py
@@ -4,7 +4,7 @@ ICON = "🌀"
 
 DESCRIPTION = f"{ICON} AI for single board computers."
 
-VERSION = "8.143.1"
+VERSION = "8.144.1"
 
 REPO_NAME = "bluer-sbc"
