+++ bluer_sbc/config.env
@@ -2,7 +2,6 @@ BLUER_SBC_CAMERA_HI_RES=1
 BLUER_SBC_CAMERA_WIDTH=728
 BLUER_SBC_CAMERA_HEIGHT=600
 BLUER_SBC_CAMERA_ROTATION=0
-BLUER_SBC_CAMERA_USE_PICAM2=0
 
 BLUER_SBC_DISPLAY_FULLSCREEN=1
