+++ bluer_sbc/tests/test_env.py
@@ -14,7 +14,6 @@ def test_blue_plugin_env():
     assert isinstance(env.BLUER_SBC_CAMERA_WIDTH, int)
     assert isinstance(env.BLUER_SBC_CAMERA_HEIGHT, int)
     assert isinstance(env.BLUER_SBC_CAMERA_ROTATION, int)
-    assert isinstance(env.BLUER_SBC_CAMERA_USE_PICAM2, int)
 
     assert isinstance(env.BLUER_SBC_DISPLAY_FULLSCREEN, bool)
